server/requirements.txt
Flask==1.1.2
Flask-Script
flask_migrate
flask-restplus
marshmallow==3.11.1
flask_marshmallow
marshmallow-sqlalchemy
psycopg2
Werkzeug==0.16.1
cerberus==1.3.3
flask-bcrypt
SQLAlchemy==1.4.7
PyJWT